requirements.txt
# Core Requirements
Flask==3.0.3
diffusers==0.28.0
transformers==4.40.1
accelerate==0.30.1
Pillow==10.3.0

# Platform-Specific PyTorch Installation
# --------------------------------------
# Choose ONE of these based on your system:

# For Apple Silicon (M1/M2/M3) Macs:
# torch==2.3.0
# torchvision==0.18.0
# torchaudio==2.3.0

# For CUDA 12.x (NVIDIA GPUs):
# torch==2.3.0+cu121
# torchvision==0.18.0+cu121
# torchaudio==2.3.0

# For CPU-only systems:
# torch==2.3.0+cpu
# torchvision==0.18.0+cpu
# torchaudio==2.3.0

# Apple Silicon Macs
pip install torch torchvision torchaudio

# NVIDIA GPUs (CUDA 12.1)
pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# CPU-only systems
pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu